[Mod] Added laser beam damager to player, added boss attack state (open laser beam);

+++ Assets/Scripts/States/Enemy_IdleState.cs
@@ -10,7 +10,7 @@ public class Enemy_IdleState : EnemyState
     {
         base.Enter();
 
-        stateTimer = enemy.useRandomIdleTime ? Random.Range((int)enemy.idleTimeRange.x, (int)enemy.idleTimeRange.y) : (int)enemy.idleTimeRange.x;
+        stateTimer = enemy.useRandomIdleTime ? Random.Range(enemy.idleTimeRange.x, enemy.idleTimeRange.y) : enemy.idleTimeRange.x;
     }
 
     public override void Update()
